# frontend/openapi.json
{
  "components": {
    "schemas": {
      "ComparePlayersRequest": {
        "properties": {
          "context": {
            "$ref": "#/components/schemas/DeterministicContext"
          },
          "left_player_id": {
            "exclusiveMinimum": 0.0,
            "title": "Left Player Id",
            "type": "integer"
          },
          "right_player_id": {
            "exclusiveMinimum": 0.0,
            "title": "Right Player Id",
            "type": "integer"
          }
        },
        "required": [
          "left_player_id",
          "right_player_id",
          "context"
        ],
        "title": "ComparePlayersRequest",
        "type": "object"
      },
      "ComparePlayersResponse": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "result": {
            "$ref": "#/components/schemas/ComparePlayersResult"
          }
        },
        "required": [
          "meta",
          "result"
        ],
        "title": "ComparePlayersResponse",
        "type": "object"
      },
      "ComparePlayersResult": {
        "properties": {
          "left_player_id": {
            "exclusiveMinimum": 0.0,
            "title": "Left Player Id",
            "type": "integer"
          },
          "metrics": {
            "additionalProperties": {
              "$ref": "#/components/schemas/MetricComparison"
            },
            "title": "Metrics",
            "type": "object"
          },
          "right_player_id": {
            "exclusiveMinimum": 0.0,
            "title": "Right Player Id",
            "type": "integer"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "left_player_id",
          "right_player_id",
          "metrics",
          "summary"
        ],
        "title": "ComparePlayersResult",
        "type": "object"
      },
      "DeterministicContext": {
        "description": "Deterministic execution contract for all stochastic endpoints.",
        "properties": {
          "data_snapshot_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._:-]+$",
            "title": "Data Snapshot Id",
            "type": "string"
          },
          "model_version": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._-]+$",
            "title": "Model Version",
            "type": "string"
          },
          "seed": {
            "maximum": 4294967295.0,
            "minimum": 0.0,
            "title": "Seed",
            "type": "integer"
          }
        },
        "required": [
          "seed",
          "model_version",
          "data_snapshot_id"
        ],
        "title": "DeterministicContext",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlayerSummary": {
        "properties": {
          "bats": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bats"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "player_id": {
            "exclusiveMinimum": 0.0,
            "title": "Player Id",
            "type": "integer"
          },
          "primary_position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Position"
          },
          "throws": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Throws"
          }
        },
        "required": [
          "player_id",
          "full_name"
        ],
        "title": "PlayerSummary",
        "type": "object"
      },
      "PredictGameRequest": {
        "properties": {
          "away_team_id": {
            "exclusiveMinimum": 0.0,
            "title": "Away Team Id",
            "type": "integer"
          },
          "context": {
            "$ref": "#/components/schemas/DeterministicContext"
          },
          "home_team_id": {
            "exclusiveMinimum": 0.0,
            "title": "Home Team Id",
            "type": "integer"
          }
        },
        "required": [
          "home_team_id",
          "away_team_id",
          "context"
        ],
        "title": "PredictGameRequest",
        "type": "object"
      },
      "PredictGameResponse": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "result": {
            "$ref": "#/components/schemas/PredictGameResult"
          }
        },
        "required": [
          "meta",
          "result"
        ],
        "title": "PredictGameResponse",
        "type": "object"
      },
      "PredictGameResult": {
        "properties": {
          "away_team_id": {
            "exclusiveMinimum": 0.0,
            "title": "Away Team Id",
            "type": "integer"
          },
          "away_win_probability": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Away Win Probability",
            "type": "number"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "explanation": {
            "items": {
              "type": "string"
            },
            "title": "Explanation",
            "type": "array"
          },
          "home_team_id": {
            "exclusiveMinimum": 0.0,
            "title": "Home Team Id",
            "type": "integer"
          },
          "home_win_probability": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Home Win Probability",
            "type": "number"
          }
        },
        "required": [
          "home_team_id",
          "away_team_id",
          "home_win_probability",
          "away_win_probability",
          "confidence",
          "explanation"
        ],
        "title": "PredictGameResult",
        "type": "object"
      },
      "ResponseMeta": {
        "properties": {
          "context": {
            "$ref": "#/components/schemas/DeterministicContext"
          },
          "generated_at_utc": {
            "format": "date-time",
            "title": "Generated At Utc",
            "type": "string"
          }
        },
        "required": [
          "context"
        ],
        "title": "ResponseMeta",
        "type": "object"
      },
      "SimulateGamePlayByPlayResponse": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "result": {
            "$ref": "#/components/schemas/SimulateGamePlayByPlayResult"
          }
        },
        "required": [
          "meta",
          "result"
        ],
        "title": "SimulateGamePlayByPlayResponse",
        "type": "object"
      },
      "SimulateGamePlayByPlayResult": {
        "properties": {
          "line_score_away": {
            "items": {
              "type": "integer"
            },
            "title": "Line Score Away",
            "type": "array"
          },
          "line_score_home": {
            "items": {
              "type": "integer"
            },
            "title": "Line Score Home",
            "type": "array"
          },
          "plays": {
            "items": {
              "$ref": "#/components/schemas/PlayByPlayEvent"
            },
            "title": "Plays",
            "type": "array"
          },
          "summary": {
            "$ref": "#/components/schemas/SimulateGameResult"
          }
        },
        "required": [
          "summary",
          "line_score_home",
          "line_score_away",
          "plays"
        ],
        "title": "SimulateGamePlayByPlayResult",
        "type": "object"
      },
      "SimulateGameRequest": {
        "properties": {
          "away_team_id": {
            "exclusiveMinimum": 0.0,
            "title": "Away Team Id",
            "type": "integer"
          },
          "context": {
            "$ref": "#/components/schemas/DeterministicContext"
          },
          "home_team_id": {
            "exclusiveMinimum": 0.0,
            "title": "Home Team Id",
            "type": "integer"
          },
          "innings": {
            "default": 9,
            "maximum": 20.0,
            "minimum": 9.0,
            "title": "Innings",
            "type": "integer"
          }
        },
        "required": [
          "home_team_id",
          "away_team_id",
          "context"
        ],
        "title": "SimulateGameRequest",
        "type": "object"
      },
      "SimulateGameResponse": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "result": {
            "$ref": "#/components/schemas/SimulateGameResult"
          }
        },
        "required": [
          "meta",
          "result"
        ],
        "title": "SimulateGameResponse",
        "type": "object"
      },
      "SimulateGameResult": {
        "properties": {
          "assumptions": {
            "items": {
              "type": "string"
            },
            "title": "Assumptions",
            "type": "array"
          },
          "away_score": {
            "minimum": 0.0,
            "title": "Away Score",
            "type": "integer"
          },
          "away_team_id": {
            "exclusiveMinimum": 0.0,
            "title": "Away Team Id",
            "type": "integer"
          },
          "home_score": {
            "minimum": 0.0,
            "title": "Home Score",
            "type": "integer"
          },
          "home_team_id": {
            "exclusiveMinimum": 0.0,
            "title": "Home Team Id",
            "type": "integer"
          },
          "innings_played": {
            "maximum": 21.0,
            "minimum": 9.0,
            "title": "Innings Played",
            "type": "integer"
          },
          "winner_team_id": {
            "exclusiveMinimum": 0.0,
            "title": "Winner Team Id",
            "type": "integer"
          }
        },
        "required": [
          "home_team_id",
          "away_team_id",
          "innings_played",
          "home_score",
          "away_score",
          "winner_team_id",
          "assumptions"
        ],
        "title": "SimulateGameResult",
        "type": "object"
      },
      "StatLeader": {
        "properties": {
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "innings_pitched": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Innings Pitched"
          },
          "plate_appearances": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plate Appearances"
          },
          "player_id": {
            "exclusiveMinimum": 0.0,
            "title": "Player Id",
            "type": "integer"
          },
          "rank": {
            "minimum": 1.0,
            "title": "Rank",
            "type": "integer"
          },
          "team_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "rank",
          "player_id",
          "full_name",
          "value"
        ],
        "title": "StatLeader",
        "type": "object"
      },
      "StatLeadersResponse": {
        "properties": {
          "direction": {
            "enum": [
              "higher_is_better",
              "lower_is_better"
            ],
            "title": "Direction",
            "type": "string"
          },
          "leaders": {
            "items": {
              "$ref": "#/components/schemas/StatLeader"
            },
            "title": "Leaders",
            "type": "array"
          },
          "metric": {
            "enum": [
              "woba",
              "wrc_plus",
              "fip",
              "k_bb_ratio"
            ],
            "title": "Metric",
            "type": "string"
          },
          "qualifier": {
            "title": "Qualifier",
            "type": "string"
          },
          "season": {
            "title": "Season",
            "type": "integer"
          }
        },
        "required": [
          "metric",
          "season",
          "direction",
          "qualifier",
          "leaders"
        ],
        "title": "StatLeadersResponse",
        "type": "object"
      },
...
      "TeamProfileFactors": {
        "description": "The seven [0, 1] matchup factors the simulator consumes.",
        "properties": {
          "command": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Command",
            "type": "number"
          },
          "discipline": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Discipline",
            "type": "number"
          },
          "offense": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Offense",
            "type": "number"
          },
          "power": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Power",
            "type": "number"
          },
          "prevention": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Prevention",
            "type": "number"
          },
          "range_factor": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Range Factor",
            "type": "number"
          },
          "speed": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Speed",
            "type": "number"
          }
        },
        "required": [
          "offense",
          "discipline",
          "power",
          "speed",
          "prevention",
          "command",
          "range_factor"
        ],
        "title": "TeamProfileFactors",
        "type": "object"
      },
      "TeamProfileResponse": {
        "properties": {
          "batters_counted": {
            "minimum": 0.0,
            "title": "Batters Counted",
            "type": "integer"
          },
          "factors": {
            "$ref": "#/components/schemas/TeamProfileFactors"
          },
          "pitchers_counted": {
            "minimum": 0.0,
            "title": "Pitchers Counted",
            "type": "integer"
          },
          "season": {
            "title": "Season",
            "type": "integer"
          },
          "source": {
            "enum": [
              "real",
              "synthetic"
            ],
            "title": "Source",
            "type": "string"
          },
          "team_fip": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Fip"
          },
          "team_id": {
            "exclusiveMinimum": 0.0,
            "title": "Team Id",
            "type": "integer"
          },
          "team_woba": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Woba"
          }
        },
        "required": [
          "team_id",
          "season",
          "source",
          "factors",
          "batters_counted",
          "pitchers_counted"
        ],
        "title": "TeamProfileResponse",
        "type": "object"
      },
      "TeamRosterResponse": {
        "properties": {
          "players": {
            "items": {
              "$ref": "#/components/schemas/PlayerSummary"
            },
            "title": "Players",
            "type": "array"
          },
          "team_id": {
            "exclusiveMinimum": 0.0,
            "title": "Team Id",
            "type": "integer"
          }
        },
        "required": [
          "team_id",
          "players"
        ],
        "title": "TeamRosterResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Deterministic sabermetrics-driven MLB simulation API.",
    "title": "baseball-sim",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/compare/players": {
      "post": {
        "operationId": "compare_players_endpoint_api_v1_compare_players_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComparePlayersRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparePlayersResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compare Players Endpoint"
      }
    },
...
    "/api/v1/players/{player_id}": {
      "get": {
        "operationId": "get_player_endpoint_api_v1_players__player_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "player_id",
            "required": true,
            "schema": {
              "title": "Player Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Player Endpoint"
      }
    },
    "/api/v1/predict/game": {
      "post": {
        "operationId": "predict_game_endpoint_api_v1_predict_game_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictGameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictGameResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Predict Game Endpoint"
      }
    },
    "/api/v1/simulate/game": {
      "post": {
        "operationId": "simulate_game_endpoint_api_v1_simulate_game_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulateGameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulateGameResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Simulate Game Endpoint"
      }
    },
    "/api/v1/simulate/game/play-by-play": {
      "post": {
        "operationId": "simulate_game_play_by_play_endpoint_api_v1_simulate_game_play_by_play_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulateGameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulateGamePlayByPlayResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Simulate Game Play By Play Endpoint"
      }
    },
    "/api/v1/stats/leaders": {
      "get": {
        "operationId": "stat_leaders_endpoint_api_v1_stats_leaders_get",
        "parameters": [
          {
            "in": "query",
            "name": "metric",
            "required": false,
            "schema": {
              "default": "woba",
              "enum": [
                "woba",
                "wrc_plus",
                "fip",
                "k_bb_ratio"
              ],
              "title": "Metric",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "season",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Season"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Playing-time qualifier (PA for hitting, IP for pitching).",
            "in": "query",
            "name": "minimum",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Playing-time qualifier (PA for hitting, IP for pitching).",
              "title": "Minimum"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatLeadersResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stat Leaders Endpoint"
      }
    },
...
    "/api/v1/teams/{team_id}/profile": {
      "get": {
        "operationId": "get_team_profile_endpoint_api_v1_teams__team_id__profile_get",
        "parameters": [
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "schema": {
              "title": "Team Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "season",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Season"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Team Profile Endpoint"
      }
    },
    "/api/v1/teams/{team_id}/roster": {
      "get": {
        "operationId": "get_team_roster_endpoint_api_v1_teams__team_id__roster_get",
        "parameters": [
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "schema": {
              "title": "Team Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamRosterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Team Roster Endpoint"
      }
    }
  }
}